+++ mayan/apps/ocr/literals.py
@@ -1,2 +1,1 @@
 DO_OCR_RETRY_DELAY = 10
-LOCK_EXPIRE = 60 * 10  # Adjust to worst case scenario