+++ mayan/apps/ocr/methods.py
@@ -5,7 +5,7 @@ from django.utils.timezone import now
 from mayan.apps.common.settings import settings_db_sync_task_delay
 
 from .events import event_ocr_document_version_submit
-from .tasks import task_do_ocr
+from .tasks import task_document_version_process
 
 
 def method_document_ocr_submit(self):
@@ -20,7 +20,7 @@ def method_document_version_ocr_submit(self):
         action_object=self.document, target=self
     )
 
-    task_do_ocr.apply_async(
+    task_document_version_process.apply_async(
         eta=now() + timedelta(seconds=settings_db_sync_task_delay.value),
-        kwargs={'document_version_pk': self.pk},
+        kwargs={'document_version_pk': self.pk}
     )