+++ mayan/apps/ocr/backends/tesseract.py
@@ -1,4 +1,5 @@
 import logging
+import os
 import shutil
 
 import sh
@@ -10,7 +11,6 @@ from mayan.apps.storage.utils import TemporaryFile
 
 from ..classes import OCRBackendBase
 from ..exceptions import OCRError
-from ..settings import setting_ocr_backend_arguments
 
 from .literals import DEFAULT_TESSERACT_BINARY_PATH, DEFAULT_TESSERACT_TIMEOUT
 
@@ -19,10 +19,10 @@ logger = logging.getLogger(name=__name__)
 
 class Tesseract(OCRBackendBase):
     def __init__(self, *args, **kwargs):
-        auto_initialize = kwargs.pop('auto_initialize', True)
-        super(Tesseract, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
+        self.read_settings()
 
-        if auto_initialize:
+        if kwargs.get('auto_initialize', True):
             self.initialize()
 
     def execute(self, *args, **kwargs):
@@ -49,7 +49,9 @@ class Tesseract(OCRBackendBase):
                 if self.language:
                     keyword_arguments['l'] = self.language
 
-                keyword_arguments['_env'] = {'OMP_THREAD_LIMIT': '1'}
+                environment = os.environ.copy()
+                environment.update(self.environment)
+                keyword_arguments['_env'] = environment
 
                 try:
 
@@ -80,8 +82,6 @@ class Tesseract(OCRBackendBase):
     def initialize(self):
         self.languages = ()
 
-        self.read_settings()
-
         try:
             self.command_tesseract = sh.Command(path=self.tesseract_binary_path)
         except sh.CommandNotFound:
@@ -110,10 +110,10 @@ class Tesseract(OCRBackendBase):
             logger.debug('Available languages: %s', ', '.join(self.languages))
 
     def read_settings(self):
-        self.tesseract_binary_path = setting_ocr_backend_arguments.value.get(
-            'tesseract_path', DEFAULT_TESSERACT_BINARY_PATH
-        )
-
-        self.command_timeout = setting_ocr_backend_arguments.value.get(
+        self.command_timeout = self.kwargs.get(
             'timeout', DEFAULT_TESSERACT_TIMEOUT
+        )
+        self.environment = self.kwargs.get('environment', {})
+        self.tesseract_binary_path = self.kwargs.get(
+            'tesseract_path', DEFAULT_TESSERACT_BINARY_PATH
         )