+++ HISTORY.rst
@@ -133,6 +133,7 @@
 
 - Remove the noopocr.NoOpOCR OCR backend.
 - Remove the pyocr OCR backend.
+- Move the ErrorLoggingMiddleware from the common app to the logging app.
 
 3.4.9 (2020-05-26)
 ==================