+++ HISTORY.rst
@@ -135,6 +135,8 @@
 - Remove the pyocr OCR backend.
 - Move the ErrorLoggingMiddleware from the common app to the logging app.
 - Allow passing environment entries to the Tesseract OCR backend.
+- Improve main menu styling and JavaScript code. Improve hover highlighting
+  and maximize space.
 
 3.4.9 (2020-05-26)
 ==================